+++ shield-status.sh
@@ -2,11 +2,16 @@
 
 ####################
 ### K.K. Ashisuto
-### VER=20240115a
+### VER=20240718a
 ####################
 
 export HOME=$(eval echo ~${SUDO_USER})
 export KUBECONFIG=${HOME}/.kube/config
+
+SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield"
+#SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield/git/develop"
+#SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield/git/feature/"
+SCRIPTS_URL_ES="https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/master/Kube/scripts"
 
 ES_PATH="$HOME/ericomshield"
 if [ ! -e $ES_PATH ];then
@@ -23,14 +28,51 @@ elif [ -f ${ES_PATH}/.es_branch ]; then
     BRANCH=$(cat ${ES_PATH}/.es_branch)
 fi
 
+if [ -f .es_version ]; then
+    S_APP_VERSION=$(cat .es_version)
+    BUILD=()
+    BUILD=(${S_APP_VERSION//./ })
+    GBUILD=${BUILD[0]}.${BUILD[1]}
+    if [[ ${BUILD[3]} ]] ;then
+        BUILD=${BUILD[2]}.${BUILD[3]}
+    else
+        BUILD=${BUILD[2]}
+    fi
+fi
+
+if [ -f .es_branch-tmp ]; then
+    BRANCH=$(cat .es_branch-tmp)
+elif [ -f ${ES_PATH}/.es_branch-tmp ]; then
+    BRANCH=$(cat ${ES_PATH}/.es_branch-tmp)
+fi
+
+if [ -f .es_version-tmp ]; then
+    S_APP_VERSION=$(cat .es_version-tmp)
+    BUILD=()
+    BUILD=(${S_APP_VERSION//./ })
+    GBUILD=${BUILD[0]}.${BUILD[1]}
+    if [[ ${BUILD[3]} ]] ;then
+        BUILD=${BUILD[2]}.${BUILD[3]}
+    else
+        BUILD=${BUILD[2]}
+    fi
+fi
+
 function usage() {
     echo "USAGE: $0 "
     exit 0
     ### for Develop only
     # [--system] [-q]
+    # status list
+    #  99 not start
+    #  
+    #  
     ##
 }
 
+rancher_ps_zero_flg=0
+rancher_ps_allactive_cnt=1
+rancher_ps_str=""
 system_flg=0
 quiet_flg=0
 args=""
@@ -50,6 +92,121 @@ do
 done
 
 
+function check_start_shield(){
+    #Shieldが開始されているか確認して変数に格納。(0の場合は、Shield停止中)
+    shield_deploy_str=`kubectl get namespaces` >/dev/null 2>&1
+    shield_deploy_common=`echo "$shield_deploy_str" | grep -c -e common`
+    shield_deploy_elk=`echo "$shield_deploy_str" | grep -c -e elk`
+    shield_deploy_farmservices=`echo "$shield_deploy_str" | grep -c -e farm-services`
+    shield_deploy_management=`echo "$shield_deploy_str" | grep -c -e management`
+    shield_deploy_proxy=`echo "$shield_deploy_str" | grep -c -e proxy`
+
+    #ステータス結果を表示
+    if { [ "$shield_deploy_common" -eq 0 ] || [ "$shield_deploy_elk" -eq 0 ] || [ "$shield_deploy_farmservices" -eq 0 ] || [ "$shield_deploy_management" -eq 0 ] || [ "$shield_deploy_proxy" -eq 0 ]; } && [ "$system_flg" -eq 0 ]; then
+        if [[ $quiet_flg -ne 1 ]]; then
+            echo "----------------------------------------------------------"
+            echo "Shield is Stopped."
+            echo "To start Shield, run ~/ericomshield/shield-start.sh"
+            echo
+            if [ $shield_deploy_common == 0 ]; then echo "Not deploy common."; fi
+            if [ $shield_deploy_elk == 0 ]; then echo "Not deploy elk."; fi
+            if [ $shield_deploy_farmservices == 0 ]; then echo "Not deploy farm-services."; fi
+            if [ $shield_deploy_management == 0 ]; then echo "Not deploy management."; fi
+            if [ $shield_deploy_proxy == 0 ]; then echo "Not deploy proxy."; fi
+            echo "----------------------------------------------------------"
+            echo "exit."
+            exit 99
+        else
+            exit 99
+        fi
+    fi
+}
+
+function check_start_system(){
+    #systemが開始されているか確認して変数に格納。(0の場合は、system停止中)
+    system_deploy_str=`kubectl get namespaces` >/dev/null 2>&1
+    system_deploy_cattle=`echo "$system_deploy_str" | grep -c -e cattle-system`
+    system_deploy_ingress=`echo "$system_deploy_str" | grep -c -e ingress-nginx`
+    system_deploy_kube=`echo "$system_deploy_str" | grep -c -e kube-system`
+
+    #ステータス結果を表示
+    if { [ "$system_deploy_cattle" -eq 0 ] || [ "$system_deploy_ingress" -eq 0 ] || [ "$system_deploy_kube" -eq 0 ]; } && [ "$system_flg" -eq 0 ]; then
+        if [[ $quiet_flg -ne 1 ]]; then
+            echo "----------------------------------------------------------"
+            echo "System is Stopped."
+            echo
+            if [ $system_deploy_cattle == 0 ]; then echo "Not deploy cattle-system."; fi
+            if [ $system_deploy_ingress == 0 ]; then echo "Not deploy ingress-nginx."; fi
+            if [ $system_deploy_kube == 0 ]; then echo "Not deploy kube-system."; fi
+            echo "----------------------------------------------------------"
+            echo "exit."
+            exit 99
+        else
+            exit 99
+        fi
+    fi
+}
+
+
+function check_count(){
+    rancher_ps_str=`rancher ps --project $(rancher projects | grep ${PROJECTNAME} | awk '{print $1}')`
+
+    #未展開の状態を検知。
+    rancher_ps_zero_flg=`echo "$rancher_ps_str" | awk '{print $4}' | grep -c -v -e STATE`
+    if [[ rancher_ps_zero_flg -eq 0 ]];then
+        if [[ $quiet_flg -ne 1 ]]; then
+            echo "Nothing active yet."
+            exit 10
+        else
+            exit 10
+        fi
+    fi
+
+    #activeステータス以外の数をカウントして変数に格納。
+    rancher_ps_not_active_cnt=`echo "$rancher_ps_str" | awk '{print $4}' | grep -c -v -e active -e STATE -e succeeded`
+}
+
+function check_login(){
+    #Rancher未ログインの場合、ログイン処理を行う。
+    rancher_login_flg=$(rancher ps 2>/dev/null | grep -c NAMESPACE) 
+    if [ $rancher_login_flg == 0 ]; then
+        rancher login --token $(cat ${ES_PATH}/.ra_apitoken) --skip-verify $(cat ${ES_PATH}/.ra_rancherurl) </dev/null >/dev/null 2>&1
+    fi
+}
+
+
+function check_status(){
+    if [[ $quiet_flg -ne 1 ]]; then
+        echo "----------------------------------------------------------"
+        echo "▼All Workload Status List▼"
+        echo
+        echo "$rancher_ps_str" | head -n 1; echo "$rancher_ps_str" | tail -n +2 | sort
+        echo "----------------------------------------------------------"
+
+        if [ $rancher_ps_not_active_cnt == 0 ]; then
+            echo "All workloads are Active !"
+            echo "----------------------------------------------------------"
+            exit 0
+        else
+            echo "▼Not Active Workload List▼"
+            echo
+            echo "$rancher_ps_str" | head -n 1; echo "$rancher_ps_str" | tail -n +2 | sort | grep -v -e active -e succeeded
+            echo "----------------------------------------------------------"
+            echo "$rancher_ps_not_active_cnt workload are not Active."
+            echo "----------------------------------------------------------"
+            exit 10
+        fi
+    else
+        if [ $rancher_ps_not_active_cnt == 0 ]; then
+            exit 0
+        else
+            exit 10
+        fi
+    fi
+
+}
+
+
 ######START#####
 
 #read ra files
@@ -62,77 +219,28 @@ else
     APITOKEN=$(cat .ra_apitoken)
 fi
 
-if [[ system_flg -ne 1 ]];then
-    PROJECTNAME="Default"
-    if [ "$BRANCH" == "Rel-19.07" ] || [ "$BRANCH" == "Rel-19.07.1" ];then
-        NAMESPACES="management proxy elk farm-services"
+check_login
+
+if [[ system_flg -eq 1 ]];then
+        PROJECTNAME="System"
+else
+    if [[ "$(echo "$BUILD < 5000" | bc)" -eq 1 ]]; then
+        ####23.05まで#######################################
+        PROJECTNAME="Default"
     else
-        NAMESPACES="management proxy elk farm-services common"
+        ####23.13以降#######################################
+        PROJECTNAME="Shield"
     fi
+fi
+
+
+if [[ system_flg -eq 1 ]];then
+    check_start_system
 else
-    PROJECTNAME="System"
-    NAMESPACES="cattle-system ingress-nginx kube-system"
+    check_start_system
+    check_start_shield
 fi
-PROJECTID=$(curl -s -k "${RANCHERURL}/v3/projects/?name=${PROJECTNAME}" \
-    -H 'content-type: application/json' \
-    -H "Authorization: Bearer $APITOKEN" \
-    | jq -r '.data[].id')
+check_count
+check_status
 
-
-
-WORKLOADS=()
-#ERR_NSs=()
-for NAMESPACE in $NAMESPACES
-do
-    WORKLOADS+=($(curl -s -k "${RANCHERURL}/v3/cluster/${CLUSTERID}/namespaces/${NAMESPACE}/yaml" \
-        -H 'content-type: application/json' \
-        -H "Authorization: Bearer $APITOKEN" \
-        | jq -c ' .items[] | [ .kind, .metadata.namespace,.metadata.name ]' 2> /dev/null | grep -v Service | grep -v ConfigMap)) || {
-        if [[ $quiet_flg -ne 1 ]]; then
-            echo "Not deploy ${NAMESPACE}."
-        fi
-        ERR_FLG=1
-        #ERR_NSs+="$NAMESPACE"
-    }
-done
-
-if [[ $ERR_FLG -eq 1 ]];then
-    if [[ $quiet_flg -ne 1 ]]; then
-        echo "exit."
-    fi
-    exit 99
-fi
-
-STATELIST=()
-for WORKLOAD in "${WORKLOADS[@]}"
-do
-        KIND=$(echo $WORKLOAD | jq -c .[0] | sed -e s/\"//g)
-        SPACE=$(echo $WORKLOAD | jq -c .[1] | sed -e s/\"//g)
-        NAME=$(echo $WORKLOAD | jq -c .[2] | sed -e s/\"//g)
-        STATELIST+=($(curl -s -k "${RANCHERURL}/v3/project/${PROJECTID}/workloads/${KIND,,}:${SPACE,,}:${NAME,,}" \
-        -H 'content-type: application/json' \
-        -H "Authorization: Bearer $APITOKEN" \
-        | jq -c '[ .namespaceId, .name, .state ] '))
-done
-
-if [[ $quiet_flg -ne 1 ]]; then
-    echo "${STATELIST[@]}" | jq -c . | grep -w "active"
-    echo ""
-fi
-NONACTIVE=$(echo "${STATELIST[@]}" | jq -c . |  grep -c -w -v "active")
-
-if [ $NONACTIVE -eq 0 ]; then
-    if [[ $quiet_flg -ne 1 ]]; then
-        echo "All workloads are Active !"
-    fi
-        exit 0
-else
-    if [[ $quiet_flg -ne 1 ]]; then
-        echo "----------------------------------------------------------"
-        echo "${STATELIST[@]}" | jq -c . |  grep -w -v "active"
-        echo "----------------------------------------------------------"
-        echo "$NONACTIVE workload are not Active."
-        echo "----------------------------------------------------------"
-    fi
-    exit 10
-fi
+exit