+++ shield-update.sh
@@ -2,7 +2,7 @@
 
 ####################
 ### K.K. Ashisuto
-### VER=20221114a
+### VER=20240718a
 ####################
 
 function usage() {
@@ -49,7 +49,9 @@ CURRENT_DIR=$(cd $(dirname $0); pwd)
 cd $CURRENT_DIR
 SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield"
 #SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield/git/develop"
+#SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield/git/feature/update-test"
 SCRIPTS_URL_ES="https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/master/Kube/scripts"
+update_flg=1
 
 if [ -f .es_branch ]; then
     BRANCH=$(cat .es_branch)
@@ -86,7 +88,7 @@ function check_ericom_user(){
             sudo mv -f ${ES_PATH}/.es_prepare ${ERICOM_PATH}/.es_prepare
             sudo chown ericom:ericom ${ERICOM_PATH}/.es_prepare
         fi
-        if [[ -f ${ES_PATH_ERICOM}/.es_prepare ]];then
+        if sudo [ -f ${ES_PATH_ERICOM}/.es_prepare ];then
             log_message "[info] Move .es_prepare flg file..."
             sudo mv -f ${ES_PATH_ERICOM}/.es_prepare ${ERICOM_PATH}/.es_prepare
             sudo chown ericom:ericom ${ERICOM_PATH}/.es_prepare
@@ -159,6 +161,7 @@ function check_args(){
 
 function select_version() {
     ### attention common setup&update&shield-prepare-servers ###
+    CRTVER=""
     CHART_VERSION=""
     VERSION_DEPLOYED=""
     if which helm >/dev/null 2>&1 ; then
@@ -173,6 +176,7 @@ function select_version() {
     echo "=================================================================="
     if [ -z $VERSION_DEPLOYED ] || [[ "$VERSION_DEPLOYED" == "request" ]] ; then
         log_message "現在インストールされているバージョン: N/A"
+        CRTVER=""
     else
         BUILD=()
         BUILD=(${VERSION_DEPLOYED//./ })
@@ -187,12 +191,13 @@ function select_version() {
             GIT_BRANCH="Rel-${GBUILD}"
         fi
         log_message "現在インストールされているバージョン: ${GIT_BRANCH}_Build:${BUILD}"
+        CRTVER="${GIT_BRANCH}_Build:${BUILD}"
     fi
     echo "=================================================================="
 
 
-    if [ -f "$ES_PREPARE" ]; then
-        log_message "実行済みのshield-prepare-serversバージョン: $(cat $ES_PREPARE)"
+    if sudo [ -f "$ES_PREPARE" ]; then
+        log_message "実行済みのshield-prepare-serversバージョン: $(sudo cat $ES_PREPARE)"
     else
         log_message "[error] shield-prepare-serversが未実行のようです。"
         echo "=================================================================="
@@ -259,6 +264,9 @@ function select_version() {
         fi
 
         echo "どのバージョンをセットアップしますか？"
+        ATTNO="0"
+        CRTNO="0"
+        TGTNO="0"
         for i in $VER
         do
             n=$(( $n + 1 ))
@@ -294,8 +302,15 @@ function select_version() {
                 if [ "$BRANCH" != "Staging" ] && [ "$BRANCH" != "Dev" ] ; then
                     if [[ $i == "eol" ]]; then
                         echo "$m: ${GIT_BRANCH}_Build:${BUILD} ※サポート終了"
+                    elif [[ $i == "attention" ]] && [[ $(basename $0) == "shield-update.sh" ]]; then
+                        echo "$m: ${GIT_BRANCH}_Build:${BUILD} "
+                        echo "======== ※これを跨いでの、shield-update.sh によるバージョンアップ不可 ========"
+                        ATTNO="$m"
                     else
                         echo "$m: ${GIT_BRANCH}_Build:${BUILD}"
+                    fi
+                    if [[ "$CRTVER" == "${GIT_BRANCH}_Build:${BUILD}" ]];then
+                        CRTNO="$m"
                     fi
                 else
                     echo "$m: Rel-$S_APP_VERSION" 
@@ -309,6 +324,7 @@ function select_version() {
             echo -n " 番号で指定してください: "
             read answer
             echo "selected versio#: $answer" >> $LOGFILE
+            TGTNO="$answer"
             if [[ -z ${vers_c[$answer]} ]] ; then
                     echo "番号が違っています。"
             else
@@ -317,6 +333,13 @@ function select_version() {
                     break
             fi
         done
+
+        if [[ "$ATTNO" -ne "0" ]] ;then
+            if [[ "$CRTNO" -gt  "$ATTNO" ]] && [[ "$TGTNO" -le  "$ATTNO" ]] ;then
+                log_message "ご指定のバージョン間でのバージョンアップはこのスクリプトではサポートされていません。"
+                fin 1
+            fi
+        fi
     fi
 
     if [ "$BRANCH" != "Staging" ] && [ "$BRANCH" != "Dev"  ]; then
@@ -372,8 +395,8 @@ function change_dir(){
 }
 
 function mv_rancher_store(){
-    if [[ $CHKBRANCH -lt 1911 ]];then
-        : 
+    if [[ $CHKBRANCH -lt 1911 ]] || [[ "$(echo "$BUILD > 5000" | bc)" -eq 1 ]];then
+        :
     else
         if [ -d ${CURRENT_DIR}/rancher-store ];then
             log_message "[start] move rancher-store"
@@ -420,6 +443,8 @@ function get_scripts() {
         curl -s -o ${ES_PATH}/delete-shield.sh -L https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/Rel-20.03/Kube/scripts/shield-setup.sh
         chmod +x ${ES_PATH}/delete-shield.sh
     fi
+    curl -s -o ${CURRENT_DIR}/shield-stop.sh -L ${SCRIPTS_URL}/shield-stop.sh
+    chmod +x ${CURRENT_DIR}/shield-stop.sh
     log_message "[end] get install scripts"
 }
 
@@ -560,8 +585,8 @@ function change_to_root(){
 
 function pre_check_prepare() {
 
-    if [ -f $ES_PREPARE ] ;then
-        PREPARE_VER=$(cat $ES_PREPARE )
+    if sudo [ -f $ES_PREPARE ] ;then
+        PREPARE_VER=$(sudo cat $ES_PREPARE )
         NOW_S_APP_VERSION=$(cat ${ES_PATH}/.es_version)
         if [[ ${PREPARE_VER} != $NOW_S_APP_VERSION ]]; then
             log_message "[info] shield-prepare was executed."
@@ -577,8 +602,8 @@ function pre_check_prepare() {
 
 function check_prepare() {
 
-    if [ -f $ES_PREPARE ] ;then
-        PREPARE_VER=$(cat $ES_PREPARE)
+    if sudo [ -f $ES_PREPARE ] ;then
+        PREPARE_VER=$(sudo cat $ES_PREPARE)
         if [[ ${PREPARE_VER} == $S_APP_VERSION ]]; then
             log_message "[info] shield-prepare was executed."
         else
@@ -686,6 +711,9 @@ else
     change_dir
     rm -f .es_update
     ./shield-stop.sh -f
+    if [ $? -ne 0 ]; then
+        fin 1
+    fi
     mv_rancher_store
     if [[ "$BRANCH" == "Rel-20.05" ]] && [[ "$OS" == "RHEL" ]]; then
         change_to_root